+++ src/index.js
@@ -647,6 +647,11 @@ const Chartkick = {
 // not ideal, but allows for simpler integration
 if (typeof window !== "undefined" && !window.Chartkick) {
   window.Chartkick = Chartkick;
+
+  // use setTimeout so charting library can come later in same JS file
+  setTimeout(function() {
+    window.dispatchEvent(new Event("chartkick:load"));
+  }, 0);
 }
 
 // backwards compatibility for esm require